Assignment4 - Digital Simulation
Implement a simple digital circuit simulator, that reads in a netlist from a file along with a stimulus set, and generates an output in a format that can be viewed in a waveform viewer. This is vaguely similar to the idea of structural Verilog code, but is much simpler and only handles combinational circuits. The details of the problem statement as well as a starter code, some examples, and a script to visualize the inputs, are provided.